+++ NEWS.md
@@ -1,3 +1,15 @@
+# TestGenerator 0.6.0
+
+* Support other database types next to duckdb: sqlserver and databricks
+
+* Remove dependency on arrow
+
+* `generateTestTables()` creates the most common used tables in the Excel sheet by default
+
+# TestGenerator 0.5.0
+
+* Support cdm version 5.4, this requires the latest version of CDMConnector
+
 # TestGenerator 0.4.0
 
 * Extra tables can be added through `readPatients.xl()`, that table then is pushed to "other tables" in the CDM reference using `patientsCDM()`.